Add flake to run readme.el without cloning the repository
The flake's default package is a script named readme that takes two
arguments. The first one is the input file (like org-readme.org), and
the second one is the output file (like README.md).

To generate a README file from a checkout of readme.el, point to the
local flake:

    nix run . -- org-readme.org README.md

To use readme.el straight from GitHub, point to
github:jeffkreeftmeijer/readme.el:

    nix run github:jeffkreeftmeijer/readme.el -- org-readme.org README.md